# docs/swagger.yaml
basePath: /v1
definitions:
...
  bands.BandResponse:
    description: Band response model
    properties:
      createdAt:
        type: string
      description:
        type: string
      genre:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  concerts.ConcertResponse:
    description: Concert response model
    properties:
      bands:
        items:
          $ref: '#/definitions/bands.BandResponse'
        type: array
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      posterUrl:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      venue:
        $ref: '#/definitions/venues.VenueResponse'
      venueId:
        type: integer
    type: object
  concerts.CreateConcertRequest:
    description: Create concert request
    properties:
      bandIds:
        items:
          type: integer
        minItems: 1
        type: array
      date:
        type: string
      description:
        maxLength: 300
        type: string
      posterUrl:
        maxLength: 100
        type: string
      title:
        maxLength: 100
        type: string
      venueId:
        type: integer
    required:
    - bandIds
    - date
    - title
    - venueId
    type: object
  concerts.ListConcertsResponse:
    description: List concerts response
    properties:
      items:
        items:
          $ref: '#/definitions/concerts.ConcertResponse'
        type: array
    type: object
  concerts.UpdateConcertRequest:
    description: Update concert request
    properties:
      bandIds:
        items:
          type: integer
        minItems: 1
        type: array
      date:
        type: string
      description:
        maxLength: 300
        type: string
      posterUrl:
        maxLength: 100
        type: string
      title:
        maxLength: 100
        type: string
      venueId:
        type: integer
    type: object
...
  venues.CreateVenueRequest:
    description: Create venue request
    properties:
      address:
        maxLength: 50
        type: string
      description:
        maxLength: 300
        type: string
      email:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - address
    - name
    type: object
  venues.ListVenuesResponse:
    description: List venues response
    properties:
      items:
        items:
          $ref: '#/definitions/venues.VenueResponse'
        type: array
    type: object
  venues.UpdateVenueRequest:
    description: Update venue request
    properties:
      address:
        maxLength: 50
        type: string
      description:
        maxLength: 300
        type: string
      email:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    type: object
  venues.VenueResponse:
    description: Venue response model
    properties:
      address:
        type: string
      createdAt:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:777
info:
  contact: {}
  description: This is a Concert booking application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Concert booking API
  version: "1.0"
paths:
...
  /admin/v1/concerts:
    get:
      description: Get a paginated list of concerts
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/concerts.ListConcertsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: List concerts
      tags:
      - Admin/Concerts
    post:
      consumes:
      - application/json
      description: Create a new concert with the provided details
      parameters:
      - description: Concert details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/concerts.CreateConcertRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/concerts.ConcertResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Create a new concert
      tags:
      - Admin/Concerts
  /admin/v1/concerts/{id}:
    delete:
      description: Delete an existing concert
      parameters:
      - description: Concert ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Delete a concert
      tags:
      - Admin/Concerts
    get:
      description: Get details of a specific concert
      parameters:
      - description: Concert ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/concerts.ConcertResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Get a concert by ID
      tags:
      - Admin/Concerts
    put:
      consumes:
      - application/json
      description: Update an existing concert
      parameters:
      - description: Concert ID
        in: path
        name: id
        required: true
        type: integer
      - description: Concert details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/concerts.UpdateConcertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/concerts.ConcertResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Update a concert
      tags:
      - Admin/Concerts
  /admin/v1/venues:
    get:
      description: Get a paginated list of venues
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/venues.ListVenuesResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: List venues
      tags:
      - Admin/Venues
    post:
      consumes:
      - application/json
      description: Create a new venue with the provided details
      parameters:
      - description: Venue details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/venues.CreateVenueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/venues.VenueResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Create a new venue
      tags:
      - Admin/Venues
  /admin/v1/venues/{id}:
    delete:
      description: Delete an existing venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Delete a venue
      tags:
      - Admin/Venues
    get:
      description: Get details of a specific venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/venues.VenueResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Get a venue by ID
      tags:
      - Admin/Venues
    put:
      consumes:
      - application/json
      description: Update an existing venue with the provided details
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: integer
      - description: Venue details to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/venues.UpdateVenueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/venues.VenueResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Update a venue
      tags:
      - Admin/Venues
...
swagger: "2.0"